ISSUE #5310 - refactor
